Create weather app
This app is generated with create-next-app. The page index.tsx is the homepage which displays the current weather at Bryggen, Bergen. The weather data is fetched from api.open-meteo.com. The file descriptions.ts contains types and translations for weather codes, as well as links to images for the different weather types.